backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis
redis[hiredis]==5.0.1

# Celery
celery[redis]==5.3.6

# Payments
razorpay==1.4.1

# WebSockets
websockets==12.0

# QR Code
qrcode[pil]==7.4.2

# Utils
httpx==0.27.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
